% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frscore-package.R
\docType{package}
\name{frscore-package}
\alias{frscore-package}
\alias{_PACKAGE}
\title{frscore: Functions for Calculating Fit-Robustness of CNA-Solutions}
\description{
Functions for automatically performing a reanalysis series on a data set using CNA, and for calculating the fit-robustness of the resulting models, as described in Parkkinen and Baumgartner (2021) \url{https://doi.org/10.1177/0049124120986200}.
}
\author{
\strong{Maintainer}: Veli-Pekka Parkkinen \email{parkkinenv@gmail.com}

Authors:
\itemize{
  \item Michael Baumgartner \email{michael.baumgartner@uib.no}
  \item Mathias Ambuehl
}

}
\keyword{internal}
