% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frscore.R
\name{frscore}
\alias{frscore}
\title{frscore}
\usage{
frscore(
  sols,
  scoretype = c("full", "supermodel", "submodel"),
  normalize = c("truemax", "idealmax", "none"),
  maxsols = 50,
  verbose = FALSE,
  print.all = FALSE
)
}
\arguments{
\item{sols}{Character vector that contains the solutions to be scored.}

\item{scoretype}{Character vector specifying the scoring method: "full"
(default, scoring is based on counting sub- and supermodels), "supermodel"
(count supermodels) "submodel" (count submodels).}

\item{normalize}{Character vector that determines the method used in
normalizing the scores. "truemax" normalizes by the actual highest score
such that the highest scoring solution types get score 1. "idealmax"
normalizes by a theoretical maximum score calculated by assuming that
all solutions of equal complexity are identical, and for every solution
of a given complexity, all solutions with lower complexity are its
submodels.}

\item{maxsols}{Integer determining the maximum number of unique solution
types found in \code{sols} to be included in the scoring.}

\item{verbose}{Logical; if \code{TRUE}, calculate and print additional
information about submodel relations among the unique solutions types found
in \code{sols}. Defaults to \code{FALSE}, which makes execution slightly
faster.}

\item{print.all}{Logical, controls the number of entries printed when
printing the results. If \code{TRUE}, results are printed as when using
\code{print.data.frame} defaults. If \code{FALSE}, 20 highest scoring
models are printed.}
}
\value{
A named list where the first element is a data frame containing
the unique solution types and their scores. Rest of the elements
contain additional information about the submodel relations among
the unique solutions types (if \code{verbose = TRUE}) and how
the function was called.
}
\description{
Calculate fit-robustness scores for a set of \code{cna}
solutions
}
\details{
\code{frscore} is intended to be used for calculating the
fit-robustness scores of \code{cna} solutions obtained by reanalyzing
a data set repeatedly, using different consistency and coverage
thresholds in each analysis.
}
\examples{
  models <- replicate(20, cna::randomCsf(4, outcome = "A"))
  results <- frscore(sols = models, verbose = TRUE)
  results

}
