% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frscored_cna.R
\name{frscored_cna}
\alias{frscored_cna}
\title{frscored_cna()}
\usage{
frscored_cna(
  x,
  fit.range = c(1, 0.7),
  granularity = 0.1,
  output = c("csf", "asf"),
  scoretype = c("full", "supermodel", "submodel"),
  normalize = c("truemax", "idealmax", "none"),
  verbose = FALSE,
  maxsols = 50,
  test.model = NULL,
  print.all = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{data.frame} or \code{configTable} to be analyzed with
\code{cna}.}

\item{fit.range}{Numeric vector of length 2; determines the maximum an
minimum values used as consistency and coverage thresholds in the
reanalysis series. Defaults to c(1, 0.7).}

\item{granularity}{Numeric scalar; consistency and coverage are varied by
this value in the reanalysis series. Default 0.1.}

\item{output}{Character vector determining whether csf's or asf's are
extracted from the results of the analyses; "csf" (default) extracts
csfs, "asf" asfs.}

\item{scoretype}{Character vector specifying the scoring method: "full"
(default, scoring is based on counting sub- and supermodels), "supermodel"
(count supermodels) "submodel" (count submodels).}

\item{normalize}{Character vector that determines the method used in
normalizing the scores. "truemax" normalizes by the actual highest score
such that the highest scoring solution types get score 1. "idealmax"
normalizes by a theoretical maximum score calculated by assuming that
all solutions of equal complexity are identical, and for every solution
of a given complexity, all solutions with lower complexity are its
submodels
.}

\item{verbose}{Logical; if \code{TRUE}, calculate and print additional
information about submodel relations among the unique solutions types found
in \code{sols}. Defaults to \code{FALSE}, which makes execution slightly
faster.}

\item{maxsols}{Integer determining the maximum number of unique solution
types found in \code{sols} to be included in the scoring.}

\item{test.model}{Character vector that specifies a single candidate
\code{cna} model whose fit-robustness score is calculated against
the results of the reanalysis series.}

\item{print.all}{Logical, controls the number of entries printed when
printing the results. If \code{TRUE}, results are printed as when using
\code{print.data.frame} defaults. If \code{FALSE}, 20 highest scoring
models are printed.}

\item{...}{Further arguments to \code{rean_cna} and \code{cna}}
}
\description{
Perform a reanalysis series on a data set and calculate the
fit-robustness scores of the resulting solution types
}
\details{
\code{frscored_cna} performs a reanalysis series on a data set with all
combinations of fit threshold values that can be generated by varying
consistency and coverage in a given range by a constant value, determined
respectively by arguments 'fit.range' and 'granularity', and calculates
fr-scores for the models returned in the analyses. If a candidate model is
provided as 'test.model', result for that model will be printed separately,
provided the model is found in the
reanalysis series, if not, the function stops.
}
\examples{
library(cna)
set.seed(6)
dat <- cna::ct2df(cna::selectCases(cna::randomCsf(5), cna::full.ct(5)))
res <- frscored_cna(dat)
res
}
