% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m_error}
\alias{m_error}
\title{Simulated data of sixteen cases with measurement error in one case}
\format{
An object of class \code{data.frame} with 16 rows and 5 columns.
}
\usage{
m_error
}
\description{
A simulated set of crisp-set configurational data
that conforms to the structure
A + B*C <-> E except for one case, 'c16',
which simulates measurement error in that case,
and including one irrelevant factor "D".
}
\keyword{datasets}
