% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_submodel_network.R
\name{plot_submodel_network}
\alias{plot_submodel_network}
\title{Visualize submodel networks}
\usage{
plot_submodel_network(
  x,
  show_clusters = TRUE,
  directed = FALSE,
  igraphlayout = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"frscore"} or \code{"frscored_cna"}.}

\item{show_clusters}{Logical; should clusters be visualized?}

\item{directed}{Logical; should submodel relations be represented as
directional?}

\item{igraphlayout}{Logical; should igraph layout be used?}

\item{...}{additional arguments to \code{visNetwork()}.}
}
\value{
A \code{visNetwork} object.
}
\description{
Visualize the output of \code{frscore()} and \code{frscored_cna()} by
generating a network representation of submodel relations among the scored
model types.
}
\details{
\code{plot_submodel_network()} takes as input a results object returned
by \code{\link[=frscore]{frscore()}} or
\code{\link[=frscored_cna]{frscored_cna()}} and uses the included adjacency
matrix of submodel relations among the scored model types to visualize
those submodel relations as a network, using
\code{\link[visNetwork:visNetwork]{visNetwork()}}. The nodes of the network
represent unique model types, and an edge between two nodes represents a
submodel relation between those model types. By default, the edges of the
network are undirected. Setting \code{directed = TRUE} produces a network
with directed edges (arrows), where the direction points from a submodel to
a supermodel. As the fr-scores of the model types depend on sub- \emph{and}
supermodel relations they have to other models, the directed network
representation provides no additional information about fr-scores over and
above the undirected network. By default, the network plot represents
clusters of model types based on edge-betweenness, calculated with
\code{\link[igraph:cluster_edge_betweenness]{cluster_edge_betweenness()}} from the
\code{igraph} package. The clusters are always base on undirected
edge-betweenness, as only the presence of submodel-relations, not their
direction, is relevant for fr-scores. The clusters can be turned off
by setting \code{show_clusters = FALSE}.
}
