% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frscored_cna.R
\name{rean_cna}
\alias{rean_cna}
\title{rean_cna}
\usage{
rean_cna(
  x,
  attempt = seq(1, 0.7, -0.1),
  ncsf = 20,
  output = c("csf", "asf"),
  ...
)
}
\arguments{
\item{attempt}{Numeric vector that contains the values from which
combinations of consistency and coverage thresholds are formed, to be used
in the analyses.}

\item{ncsf}{Scalar integer, determines the maximum number of csfs calculated
in a each \code{cna} analysis, if \code{output = "csf"}. Defaults to 20.}

\item{output}{Character vector that determines whether csfs or asfs are
returned; \code{"csf"} (default) returns csfs \code{"asf"} asfs.}

\item{...}{Any arguments to \code{cna} except \code{con}, \code{cov} or
\code{con.msc}, at a minimum the arxument \code{x}, a \code{data.frame} or
a \code{configTable}. The effect of argument \code{what} will be overriden
by \code{output}.}
}
\value{
A list, where each element is a data frame containing the results of
a single analysis of the input data set with \code{cna}, each using a
different combination of consistency and coverage threshold values. These
values are added to the output as extra columns 'cnacon' and 'cnacov'.
}
\description{
Perform a reanalysis series on a data set with \code{cna} using
all combinations of consistency and coverage threshold values in a given
range of values
}
\details{
\code{rean_cna} performs a reanalysis series based on which
fr-scores can be calculated output = "asf" returns asfs, output = "csf"
csfs.
}
