% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_submodel_network.R
\name{submodel_adjacencies_to_igraph}
\alias{submodel_adjacencies_to_igraph}
\title{Generate igraph from submodel adjacencies}
\usage{
submodel_adjacencies_to_igraph(x)
}
\arguments{
\item{x}{An object of class \code{"frscore"} or \code{"frscored_cna"}.}
}
\value{
An \code{igraph} graph.
}
\description{
Generate igraph from submodel adjacencies
}
